+++ packages/matter-node-ble.js/src/ble/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @license
+ * Copyright 2022-2023 Project CHIP Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from "./BluetoothNode.js";
+export * from "./BleBroadcaster.js";
+export * from "./BleNetworkInterface.js";